backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.2.0

# Database & Authentication (WORKING COMBINATION)
supabase==2.4.0
httpx==0.24.1
gotrue==2.4.4
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Computer Vision & Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0

# Scientific Computing & ML
scikit-learn==1.3.2
tensorflow==2.14.0
matplotlib==3.8.2
seaborn==0.13.0

# Data Processing
pandas==2.1.3

# Utilities
requests==2.31.0
tqdm==4.66.1
